+++ lib/webserver/apihandlers.py
@@ -60,6 +60,7 @@ class ApiHandlers:
 
         Query args:
             fields (String, comma seperated): list of fields to return
+            byName (string): 1, 0 or true, false
         """
         series_name = unquote(request.match_info['series_name'])
         by_name = False
@@ -86,8 +87,14 @@ class ApiHandlers:
 
         Query args:
             fields (String, comma seperated): list of fields to return
+            byName (string): 1, 0 or true, false
         """
         rid = unquote(request.match_info['rid'])
+        by_name = False
+        if 'byName' in request.rel_url.query:
+            q_by_name = request.rel_url.query['byName']
+            if q_by_name == "1" or q_by_name.lower() == "true":
+                by_name = True
         only_future = False
         if "forecastFutureOnly" in request.rel_url.query:
             val = request.rel_url.query['forecastFutureOnly']
@@ -103,7 +110,7 @@ class ApiHandlers:
         return web.json_response(
             data=await BaseHandler.resp_get_all_series_output(
                 rid, fields=fields, forecast_future_only=only_future,
-                types=types),
+                types=types, by_name=by_name),
             dumps=safe_json_dumps)
 
     @classmethod
@@ -113,10 +120,19 @@ class ApiHandlers:
 
         Returns:
             _type_: _description_
+
+        Query args:
+            byName (string): 1, 0 or true, false
         """
-        rid = request.match_info['rid']
+        rid = unquote(request.match_info['rid'])
+        by_name = False
+        if 'byName' in request.rel_url.query:
+            q_by_name = request.rel_url.query['byName']
+            if q_by_name == "1" or q_by_name.lower() == "true":
+                by_name = True
         job_name = unquote(request.match_info['job_config_name'])
-        await BaseHandler.resp_resolve_series_job_status(rid, job_name)
+        await BaseHandler.resp_resolve_series_job_status(rid, job_name,
+                                                         by_name=by_name)
         return web.json_response(
             {}, dumps=safe_json_dumps, status=200)
 
@@ -170,10 +186,20 @@ class ApiHandlers:
 
         Returns:
             _type_: _description_
+
+        Query args:
+            byName (string): 1, 0 or true, false
         """
-        rid = request.match_info['rid']
+        rid = unquote(request.match_info['rid'])
+        by_name = False
+        if 'byName' in request.rel_url.query:
+            q_by_name = request.rel_url.query['byName']
+            if q_by_name == "1" or q_by_name.lower() == "true":
+                by_name = True
         return web.json_response(
-            data={}, status=await BaseHandler.resp_remove_series(rid))
+            data={}, status=await BaseHandler.resp_remove_series(
+                rid,
+                by_name=by_name))
 
     @classmethod
     @EnodoAuth.auth.required
@@ -207,12 +233,20 @@ class ApiHandlers:
 
         Returns:
             _type_: _description_
+
+        Query args:
+            byName (string): 1, 0 or true, false
         """
-        rid = request.match_info['rid']
+        rid = unquote(request.match_info['rid'])
+        by_name = False
+        if 'byName' in request.rel_url.query:
+            q_by_name = request.rel_url.query['byName']
+            if q_by_name == "1" or q_by_name.lower() == "true":
+                by_name = True
         job_config_name = urllib.parse.unquote(
             request.match_info['job_config_name'])
         data, status = await BaseHandler.resp_remove_job_config(
-            rid, job_config_name)
+            rid, job_config_name, by_name=by_name)
         return web.json_response(
             data=data, status=status)